move search action to thunks file, reconcile file naming
